Add overhead experiment with nearly unmodified ReentrantLock
The problem may happen when running this experiment many times
with the script: sh repeat.sh 30000 sh run_experiment.sh 200 10 s 1000

It runs 30000 times the experiment in which safe.ReentrantLock is used
to protect each of the 10 counters that will be incremented by 200 threads.
Each thread will handle only one lock and this is evenly distributed, so each
lock will have 20 threads incrementing it concurrently. Each thread will increment
the counter 1000 times, so in the end, each counter should have the count 20000 and
the sum of all counters should be 200000 which is finally printed to stdout.

The problem seems to be a missing unpark signal and some remaining threads
keep waiting for an unlocked resource.

If you change the script to use default ReentrantLock (replace 's' for 'r'),
then this problem no longer shows (at least not until now).